Perform GC write barriers for shared PMCs only on the main interp
Performing write barriers on a different interp than the one owning a
PMC leads to corruption of the GC's objects and dirty lists since the
write barrier removes the PMC from the objects list of the generation
the PMC belongs to. This causes the next_free pointer of the main
thread's GC's objects list to point to a place in the child interp's
objects list.

Fixed by checking the shared flag and delegating the call to the
main_interp if the current interp is a child.